Cleaning up main
